Add frequency ramp for BM1397 (#1069)
* Accept and store float frequency values

* Unify PLL resolver for 1366/1368/1370

Pick the closest frequency to the requested target. If there are multiple PLL settings, pick the one with first the lowest VDO frequency, and finally the lowest posdividers.

Merged frequency ramp start and loop. This solves overshooting when requesting a target frequency within the current step size. It also eliminates the duplicate setting of the target frequency is that's on a step boundary.

Cleaned up several unused defines.

* Restore postdiv2 < postdiv1 condition for BM1368 and BM1370

* Whitespace

* Clean up intermediate functions

Simplify frequency_transition further, finally fix overshoot

* Code cleanup

* Add debug logging for BM1397

* Fix log

* Experimental PLL BM1397 support

* Flip postdiv condition for readability

* Add test

* Add asic project to test components

* Add frequency ramp at start for BM1397

* Remove asic module from test build

* Missing import

* Wrong hash function call

* Remove bm1397 module for test

* Add missing import

* Try to fix test_job_command

* Ignore test for now, can't get it to compile

* Fix BM1370 fb_range

* Add bunch of test cases

* Revert test_pll.c

* Whitespace

* Fix test